Replace string literals with enum members
